+++ packages/Adapter-SQLite3/lib/index.js
@@ -21,7 +21,7 @@ var SQLite = /** @class */ (function (_super) {
         var _this = _super.call(this, false, true) || this;
         _this.plugin = {
             name: "SQLite Adapter",
-            version: 2.04
+            version: 2.05
         };
         _this._ai = {};
         _this._query = _this._query.bind(_this);