+++ packages/Adapter-SQLite3/src/index.ts
@@ -10,7 +10,7 @@ export class SQLite  extends nanoSQLMemoryIndex {
 
     plugin: InanoSQLPlugin = {
         name: "SQLite Adapter",
-        version: 2.04
+        version: 2.05
     };
 
     nSQL: InanoSQLInstance;